Move FormRequests to FormRequests directory
