+++ app/Domain/Integrations/Controllers/IntegrationController.php
@@ -9,6 +9,10 @@ use App\Domain\Auth\CurrentUser;
 use App\Domain\Contacts\Contact;
 use App\Domain\Contacts\ContactType;
 use App\Domain\Contacts\Repositories\ContactRepository;
+use App\Domain\Integrations\FormRequests\StoreIntegration;
+use App\Domain\Integrations\FormRequests\UpdateBasicInfo;
+use App\Domain\Integrations\FormRequests\UpdateBillingInfo;
+use App\Domain\Integrations\FormRequests\UpdateContactInfo;
 use App\Domain\Integrations\IntegrationType;
 use App\Domain\Integrations\Mappers\StoreIntegrationMapper;
 use App\Domain\Integrations\Repositories\IntegrationRepository;
